added sudo support with no password

+++ push.sh
@@ -1,4 +1,5 @@
-docker login --username tanvirtin
-docker build --rm -f Dockerfile -t tindev .
-docker tag tindev tanvirtin/tindev
-docker push tanvirtin/tindev
+if docker login --username tanvirtin; then
+    docker build --rm -f Dockerfile -t tindev .
+    docker tag tindev tanvirtin/tindev
+    docker push tanvirtin/tindev
+fi
